feat: 그룹 요청 리스트 받기 , 수락, 거절 기능

+++ frontend/ARchive/app/src/main/java/com/droidblossom/archive/data/dto/group/response/GroupInviteResponseDto.kt
@@ -6,7 +6,7 @@ import com.droidblossom.archive.domain.model.group.GroupPage
 data class GroupInviteResponseDto(
     val createdAt: String,
     val description: String,
-    val groupId: Int,
+    val groupId: Long,
     val groupName: String,
     val groupOwnerName: String,
     val groupProfileUrl: String
